requirements-monitor.txt
requests>=2.31.0
PyYAML>=6.0
APScheduler>=3.10.0
beautifulsoup4>=4.12.0
